Remove table stripes, add border edges, fix row-actions with table-actions=false
