+++ components/SortableTable/selection.js
@@ -132,10 +132,9 @@ export default {
         return;
       }
 
-      const check = tgtRow.find('label.selection-checkbox');
-      const nodeId = check.data('node-id');
+      const nodeId = tgtRow.data('node-id');
 
-      if ( !nodeId || !check || !check.length || check[0].disabled ) {
+      if ( !nodeId ) {
         return;
       }
