--- models/cluster.js
@@ -1,55 +1,0 @@
-import { insertAt } from '@/utils/array';
-
-export default {
-
-  _availableActions() {
-    const out = this._standardActions;
-
-    insertAt(out, 0, {
-      action:     'openShell',
-      label:      'Kubectl Shell',
-      icon:       'icon icon-terminal',
-      enabled:    !!this.links.shell,
-    });
-
-    return out;
-  },
-
-  openShell() {
-    return () => {
-      this.$dispatch('wm/open', {
-        id:        `kubectl-${ this.id }`,
-        label:     this.$rootGetters['i18n/t']('wm.kubectlShell.title', { name: this.nameDisplay }),
-        icon:      'terminal',
-        component: 'KubectlShell',
-        attrs:     {
-          cluster: this,
-          pod:     {}
-        }
-      }, { root: true });
-    };
-  },
-
-  isReady() {
-    return this.hasCondition('Ready');
-  },
-
-  configName() {
-    const allKeys = Object.keys(this.spec);
-    const configKey = allKeys.find( kee => kee.includes('Config'));
-
-    return configKey;
-  },
-
-  kubernetesVersion() {
-    if ( this?.status?.version?.gitVersion ) {
-      return this.status.version.gitVersion;
-    } else {
-      return this.$rootGetters['i18n/t']('generic.unknown');
-    }
-  },
-
-  canDelete() {
-    return this.hasLink('remove') && !this?.spec?.internal;
-  },
-};
